Update home page modal content
The text within the home page modal has been updated to promote a free demo instead of just inviting users to get in touch. The new text also details the simplicity of booking the demo and outlines what they will learn during the session. This should increase the effectiveness of capturing potential customers' interest and encouraging them to engage with the service.