Replace button links with <x-form-cta/> and create a new component
This commit replaces older standard links with the <x-form-cta/> component on the front page. This component allows for better reusability and design consistency. A new component, 'business-landing-page.blade.php' is also created, implementing a new design structure, which includes the <x-form-cta/> calls. This promotes better code organization and enhances the user interface.