Update layout adjustments and modify monthly order count to yearly
The commit contains modifications to the layout of the home template and form CTA component. Furthermore, there are changes to the PricesService. These changes adapt the display from monthly order counts to yearly. The template revisions improve responsiveness of the display, and the changes to PricesService provide more accurate information about the order capabilities of the service plans.